+++ Examples/YandexMobileAdsExample/YandexMobileAdsExample/Third-PartyMediation/IronSource/IronSourceInterstitialViewController.swift
@@ -28,6 +28,7 @@ class IronSourceInterstitialViewController: UIViewController {
 
     @IBAction func showAd(_ sender: UIButton) {
         IronSource.showInterstitial(with: self)
+        showButton.isEnabled = false
     }
 }