+++ changelog/adapter/admob-mobileads/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 6.0.0.0
+
+### Updated
+* Added support for Yandex Mobile Ads SDK 6.0.0
+* Added support for Google Mobile Ads SDK version 10.9.0
+* Updated minimum supported Google Mobile Ads SDK version to 10.9.0
+
 ## Version 5.9.1.0
 
 ### Updated