+++ changelog/mediation/mobileads-ironsource/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## Version 7.4.0.0
+
+#### Updated
+* Added support for IronSource SDK version 7.4.0.0
+* Updated minimum supported IronSource SDK version to 7.4.0.0
+* Updated minimum supported Yandex Mobile Ads SDK version to 6.0.0
+
 ## Version 7.3.0.3
 
 #### Updated